move utils, update comments
