+++ apps/src/aichat/views/ChatWorkspace.tsx
@@ -11,7 +11,7 @@ import React, {
   useImperativeHandle,
 } from 'react';
 
-import {findChangedProperties} from '@cdo/apps/aichatLab/redux/utils';
+import {findChangedProperties} from '@cdo/apps/aichat/redux/utils';
 import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
 import usePrevious from '@cdo/apps/util/usePrevious';
 import {AiChatClientTypes} from '@cdo/generated-scripts/sharedConstants';
@@ -256,7 +256,6 @@ const ChatWorkspace = forwardRef<ChatWorkspaceHandle, ChatWorkspaceProps>(
     // Whenever model parameters change, 1) reset the chat session if necessary,
     // and 2) log the changed properties to the chat history.
     const previousParameters: ModelParameters = usePrevious(modelParameters);
-    // FIX!!
     useEffect(() => {
       const changedProperties = findChangedProperties(
         previousParameters,
