+++ apps/src/aichatLab/redux/thunks/onSaveComplete.ts
@@ -1,4 +1,5 @@
 import {sendAnalytics} from '@cdo/apps/aichat/redux';
+import {findChangedProperties} from '@cdo/apps/aichat/redux/utils';
 import {SaveType, ViewMode} from '@cdo/apps/aichatLab/types';
 import {sendProgressReport} from '@cdo/apps/code-studio/progressRedux';
 import {TestResults} from '@cdo/apps/constants';
@@ -7,7 +8,6 @@ import {RootState} from '@cdo/apps/types/redux';
 import {AppDispatch} from '@cdo/apps/util/reduxHooks';
 
 import {endSave, setSavedAiCustomizations, setViewMode} from '../slice';
-import {findChangedProperties} from '../utils';
 
 const saveTypeToAnalyticsEvent: {[key in SaveType]: string} = {
   updateChatbot: EVENTS.UPDATE_CHATBOT,
