+++ apps/test/unit/aichat/redux/utilsTest.ts
@@ -1,4 +1,4 @@
-import * as utils from '@cdo/apps/aichatLab/redux/utils';
+import * as utils from '@cdo/apps/aichat/redux/utils';
 import {AiCustomizations} from '@cdo/apps/aichatLab/types';
 import {EMPTY_MODEL_CARD_INFO} from '@cdo/apps/aichatLab/views/modelCustomization/constants';
 import {AiChatModelIds} from '@cdo/generated-scripts/sharedConstants';
